seps-plugin/scripts/engine/requirements.txt
psd-tools>=1.9.28
Pillow>=10.0.0
numpy>=1.24.0
